mod:: OptionModal 컴포넌트 분리 - timetable tsx 내에 있던 OptionModal 컴포넌트를 다른 파일로 분리했습니다
